+++ crypto/aes/cts/cts_test.go
@@ -1,4 +1,4 @@
-package aescts
+package cts
 
 import (
 	"bytes"
